Move deprecated argument test
